requirements.txt
requests>=2.24.0
httplib2>=0.18.1
pymssql>=2.1.4
coc.py>=0.3.4
asyncpg>=0.20.1
loguru>=0.5.1
discord.py
GitPython
google-api-python-client>=1.9.3
oauth2client>=4.1.3
asyncio>=3.4.3
Pillow>=7.1.2
PyYAML>=5.3.1
aiohttp>=3.6.2